refactor: extract AiCardTab subcomponent from CardFactoryModal.tsx (#8608) [part 6]
Move the AI Create tab content (Tier-1 declarative + Tier-2 custom-code
generation flows) into its own file, `cardFactoryAiTab.tsx`. The component
was previously defined inline at the bottom of CardFactoryModal.tsx and
is only consumed in one place via the existing tab JSX.

This removes a number of imports from CardFactoryModal.tsx that were only
used by the inline subcomponent (AiGenerationPanel, T1Preview/T2Preview,
CARD_T1_SYSTEM_PROMPT/CARD_T2_SYSTEM_PROMPT, validateT1Result/
validateT2Result, AiCardT1Result/AiCardT2Result, AiMode), keeping the
parent file's import surface focused on the modal shell + declarative/
code-tab logic.

Additionally promotes two magic numbers in the extracted component to
named constants (DEFAULT_T1_LIMIT, DEFAULT_CARD_WIDTH).

Zero behavior change: the AiCardTab is invoked exactly the same way
from the parent modal, with the same `onCardCreated` callback prop.

Reduces CardFactoryModal.tsx from 1356 to 1238 LOC (-118).
